+++ managers.d.ts
