+++ rollup.config.js
@@ -1,19 +1,24 @@
-import babel from "@rollup/plugin-babel";
-import commonjs from "@rollup/plugin-commonjs";
-import resolve from "@rollup/plugin-node-resolve";
-import replace from "@rollup/plugin-replace";
-import alias from "@rollup/plugin-alias";
-import svgr from "@svgr/rollup";
-import styles from "rollup-plugin-styles";
-import json from "@rollup/plugin-json";
-import peerDepsExternal from "rollup-plugin-peer-deps-external";
-import { mergeDeepLeft } from "ramda";
-import copy from "rollup-plugin-copy";
-import packageJson from "./package.json";
 import path from "path";
 
+import alias from "@rollup/plugin-alias";
+import babel from "@rollup/plugin-babel";
+import commonjs from "@rollup/plugin-commonjs";
+import json from "@rollup/plugin-json";
+import resolve from "@rollup/plugin-node-resolve";
+import replace from "@rollup/plugin-replace";
+import svgr from "@svgr/rollup";
+import { mergeDeepLeft } from "ramda";
+import cleaner from "rollup-plugin-cleaner";
+import copy from "rollup-plugin-copy";
+import peerDepsExternal from "rollup-plugin-peer-deps-external";
+import styles from "rollup-plugin-styles";
+
+import packageJson from "./package.json";
+
 const commonResolve = require("@bigbinary/neeto-commons-frontend/configs/nanos/webpack/resolve.js");
+
 const projectResolve = require("./resolve.js");
+
 const { alias: aliasEntries } = mergeDeepLeft(projectResolve, commonResolve);
 
 const config = args => {
@@ -22,6 +27,9 @@ const config = args => {
     : __dirname;
 
   const plugins = [
+    cleaner({
+      targets: [path.resolve(__dirname, `${destination}dist/`)],
+    }),
     peerDepsExternal(),
     alias({ entries: aliasEntries }),
     replace({
@@ -29,10 +37,7 @@ const config = args => {
       preventAssignment: true,
     }),
     svgr(),
-    babel({
-      exclude: "node_modules/**",
-      babelHelpers: "runtime",
-    }),
+    babel({ exclude: "node_modules/**", babelHelpers: "runtime" }),
     resolve({
       preferBuiltins: true,
       extensions: [".js", ".jsx", ".svg"],
@@ -40,69 +45,36 @@ const config = args => {
     }),
     commonjs({ include: /\**node_modules\**/ }),
     json(),
+    styles({
+      extensions: [".css", ".scss", ".min.css"],
+      mode: ["extract", "index.css"],
+    }),
+    args.app &&
+      copy({
+        targets: [
+          { src: "package.json", dest: destination },
+          { src: "types/", dest: destination },
+        ],
+      }),
   ];
 
-  const formats = ["esm", "cjs"];
+  const input = {
+    index: "./src/components/index.js",
+    layouts: "./src/components/layouts/index.js",
+    formik: "./src/components/formik/index.js",
+    managers: "./src/managers/index.js",
+  };
 
-  const getOutputFileName = (name, format) =>
-    format === "esm" ? `${name}.js` : `${name}.cjs.js`;
-
-  return [
-    {
-      input: "./src/components/index.js",
-      output: formats.map(format => ({
-        file: path.resolve(destination, getOutputFileName("index", format)),
-        format,
-        sourcemap: true,
-        assetFileNames: "[name][extname]",
-      })),
-      external: ["@bigbinary/neetoui/managers"],
-      plugins: [
-        ...plugins,
-        styles({
-          extensions: [".css", ".scss", ".min.css"],
-          mode: ["extract", "index.css"],
-        }),
-        args.app &&
-          copy({
-            targets: [
-              { src: "package.json", dest: destination },
-              { src: "index.d.ts", dest: destination },
-              { src: "formik.d.ts", dest: destination },
-              { src: "managers.d.ts", dest: destination },
-            ],
-          }),
-      ].filter(Boolean),
-    },
-    {
-      input: "./src/components/layouts/index.js",
-      output: formats.map(format => ({
-        file: path.resolve(destination, getOutputFileName("layouts", format)),
-        format,
-        sourcemap: true,
-      })),
-      plugins,
-    },
-    {
-      input: "./src/components/formik/index.js",
-      output: formats.map(format => ({
-        file: getOutputFileName("formik", format),
-        format,
-        sourcemap: true,
-      })),
-      external: ["@bigbinary/neetoui/managers"],
-      plugins,
-    },
-    {
-      input: "./src/managers/index.js",
-      output: formats.map(format => ({
-        file: path.resolve(destination, getOutputFileName("managers", format)),
-        format,
-        sourcemap: true,
-      })),
-      plugins,
-    },
-  ];
+  return {
+    input,
+    output: ["esm", "cjs"].map(format => ({
+      dir: path.resolve(destination, `dist/${format === "cjs" ? "cjs/" : ""}`),
+      format,
+      sourcemap: true,
+      exports: "auto",
+    })),
+    plugins,
+  };
 };
 
 export default config;
